pull out setUp into mixin
